Memory

+++ Memory/Signatures.cs
@@ -65,7 +65,7 @@ namespace Memory
         }
         public (byte, bool)[] sig { get; private set; } = null;
         public string txt = "No_Signatures";
-        public bool Test(byte a, int id)
+        public bool Test(byte a, long id)
         {
             if (sig[id].Item2) { return a == sig[id].Item1; }
             return true;
